refactored useCanvas
